refactor: float button - 스크롤 다운 시 버튼 사라지고, 업될때 다시 생기도록 함
