+++ apps/web/components/Buttons/index.module.scss
@@ -58,11 +58,26 @@
 .float {
   width: 100px;
   padding: 3px;
-
   position: fixed;
   bottom: 30px;
 
   z-index: 80;
-
   @include m.borderStyles();
+
+  transition: transform 0.3s ease;
+}
+
+.floatBtnBox {
+  transition:
+    opacity 0.3s ease,
+    visibility 0.3s ease;
+}
+
+.hide {
+  opacity: 0;
+  visibility: hidden;
+}
+.show {
+  opacity: 1;
+  visibility: visible;
 }
